Add a longnostic to a dataframe

Match a longnostic to the id of a dataframe. This saves you creating a longnostic and then joining it back to a dataframe

add_l_max(data, id, var)

add_l_diff(data, id, var, lag = 1)

add_l_mean(data, id, var)

add_l_median(data, id, var)

add_l_min(data, id, var)

add_l_n_obs(data, id)

add_l_q1(data, id, var)

add_l_q3(data, id, var)

add_l_sd(data, id, var)

add_l_slope(data, id, formula)

Arguments

data
data.frame to explore

id
vector of ids to define which values belong to which individual

var
vector of values for individuals, needs to match the id vector

lag
the lag to use, default to 1 (used with l_diff)

formula
character, a formula representing the slope of interest (used in l_slope)

Examples

add_l_max(data = wages, id = id, var = lnw)

#> # A tibble: 6,402 x 10
#>       id l_max   lnw exper   ged postexp black hispanic   hgc uerate
#>    <int> <dbl> <dbl> <dbl> <int>   <dbl> <int>    <int> <int>  <dbl>
#>  1    31  2.13  1.49 0.015     1   0.015     0        1     8   3.21
#>  2    31  2.13  1.43 0.715     1   0.715     0        1     8   3.21
#>  3    31  2.13  1.47 1.73      1   1.73      0        1     8   3.21
#>  4    31  2.13  1.75 2.77      1   2.77      0        1     8   3.3 
#>  5    31  2.13  1.93 3.93      1   3.93      0        1     8   2.89
#>  6    31  2.13  1.71 4.95      1   4.95      0        1     8   2.49
#>  7    31  2.13  2.09 5.96      1   5.96      0        1     8   2.6 
#>  8    31  2.13  2.13 6.98      1   6.98      0        1     8   4.8 
#>  9    36  2.93  1.98 0.315     1   0.315     0        0     9   4.89
#> 10    36  2.93  1.80 0.983     1   0.983     0        0     9   7.4 
#> # … with 6,392 more rows

add_l_diff(data = wages, id = id, var = lnw)

#> # A tibble: 6,402 x 10
#>       id l_diff_1   lnw exper   ged postexp black hispanic   hgc uerate
#>    <int>    <dbl> <dbl> <dbl> <int>   <dbl> <int>    <int> <int>  <dbl>
#>  1    31    0.377  1.49 0.015     1   0.015     0        1     8   3.21
#>  2    31    0.377  1.43 0.715     1   0.715     0        1     8   3.21
#>  3    31    0.377  1.47 1.73      1   1.73      0        1     8   3.21
#>  4    31    0.377  1.75 2.77      1   2.77      0        1     8   3.3 
#>  5    31    0.377  1.93 3.93      1   3.93      0        1     8   2.89
#>  6    31    0.377  1.71 4.95      1   4.95      0        1     8   2.49
#>  7    31    0.377  2.09 5.96      1   5.96      0        1     8   2.6 
#>  8    31    0.377  2.13 6.98      1   6.98      0        1     8   4.8 
#>  9    36    1.11   1.98 0.315     1   0.315     0        0     9   4.89
#> 10    36    1.11   1.80 0.983     1   0.983     0        0     9   7.4 
#> # … with 6,392 more rows

add_l_diff(data = wages, id = id, var = lnw, lag = 2)

#> # A tibble: 6,402 x 10
#>       id l_diff_2   lnw exper   ged postexp black hispanic   hgc uerate
#>    <int>    <dbl> <dbl> <dbl> <int>   <dbl> <int>    <int> <int>  <dbl>
#>  1    31    0.462  1.49 0.015     1   0.015     0        1     8   3.21
#>  2    31    0.462  1.43 0.715     1   0.715     0        1     8   3.21
#>  3    31    0.462  1.47 1.73      1   1.73      0        1     8   3.21
#>  4    31    0.462  1.75 2.77      1   2.77      0        1     8   3.3 
#>  5    31    0.462  1.93 3.93      1   3.93      0        1     8   2.89
#>  6    31    0.462  1.71 4.95      1   4.95      0        1     8   2.49
#>  7    31    0.462  2.09 5.96      1   5.96      0        1     8   2.6 
#>  8    31    0.462  2.13 6.98      1   6.98      0        1     8   4.8 
#>  9    36    0.775  1.98 0.315     1   0.315     0        0     9   4.89
#> 10    36    0.775  1.80 0.983     1   0.983     0        0     9   7.4 
#> # … with 6,392 more rows