Filter by the number of observations for a key.

When exploring longitudinal data it can be useful to filter by the number of observations in a compact way. filter_n_obs allows for the user to filter by the number of observations for each key.

filter_n_obs(.data, filter, ...)

# S3 method for tbl_ts
filter_n_obs(.data, filter, ...)

Arguments

.data
data.frame

filter
A description of how you want to filter the number of observations for each key, in terms of n_obs. See examples for more detail.

...
additional arguments

Value

data.frame filtered by the number of observations, with an additional column n_obs, which contains the number of observations for each key.

Examples

wages %>% filter_n_obs(n_obs > 10)

#> # A tsibble: 1,105 x 10 [!]
#> # Key:       id [97]
#>       id    xp n_obs ln_wages   ged xp_since_ged black hispanic high_grade
#>    <int> <dbl> <int>    <dbl> <int>        <dbl> <int>    <int>      <int>
#>  1   134 0.192    12     2.00     0         0        0        0         11
#>  2   134 0.972    12     2.10     0         0        0        0         11
#>  3   134 1.93     12     2.32     0         0        0        0         11
#>  4   134 2.95     12     2.39     0         0        0        0         11
#>  5   134 3.92     12     2.26     1         0        0        0         11
#>  6   134 4.64     12     2.33     1         0.72     0        0         11
#>  7   134 5.64     12     2.30     1         1.72     0        0         11
#>  8   134 6.75     12     2.67     1         2.84     0        0         11
#>  9   134 8.25     12     2.81     1         4.34     0        0         11
#> 10   134 9.02     12     2.93     1         5.11     0        0         11
#> # … with 1,095 more rows, and 1 more variable: unemploy_rate <dbl>

wages %>% filter_n_obs(n_obs == 2)

#> # A tsibble: 78 x 10 [!]
#> # Key:       id [39]
#>       id    xp n_obs ln_wages   ged xp_since_ged black hispanic high_grade
#>    <int> <dbl> <int>    <dbl> <int>        <dbl> <int>    <int>      <int>
#>  1   518 0.525     2    1.27      1        0.525     1        0          8
#>  2   518 1.93      2    1.61      1        1.93      1        0          8
#>  3  1687 0.688     2    1.66      0        0         1        0         10
#>  4  1687 0.825     2    1.26      0        0         1        0         10
#>  5  2178 0.923     2    1.52      0        0         0        0          8
#>  6  2178 1.96      2    1.75      0        0         0        0          8
#>  7  2456 4         2    0.898     0        0         0        0         11
#>  8  2456 5         2    1.22      0        0         0        0         11
#>  9  2612 0.875     2    1.42      1        0.875     0        0          8
#> 10  2612 1.14      2    1.57      1        1.14      0        0          8
#> # … with 68 more rows, and 1 more variable: unemploy_rate <dbl>