Return keys nearest to a given statistics or summary.

Return keys nearest to a given statistics or summary.

keys_near(.data, var, top_n = 1, funs = l_five_num, ...)

Arguments

.data
data.frame

var
variable to summarise

top_n
top number of closest observations to return - default is 1, which will also return ties.

funs
named list of functions to summarise by. Default is a given list of the five number summary, l_five_num.

...
extra arguments to pass to mutate_at when performing the summary as given by funs.

key
key, which identifies unique observations.

Value

data.frame containing keys closest to a given statistic.

Examples

wages %>%
  key_slope(ln_wages ~ xp) %>%
  keys_near(key = id,
            var = .slope_xp)

#> Warning: Outer names are only allowed for unnamed scalar atomic inputs

#> # A tibble: 6 x 5
#>      id .slope_xp stat  stat_value stat_diff
#>   <int>     <dbl> <fct>      <dbl>     <dbl>
#> 1  7918  -4.58    min     -4.58    0        
#> 2  2594  -0.00768 q_25    -0.00769 0.0000127
#> 3  2305   0.0480  med      0.0480  0.0000498
#> 4 10380   0.0479  med      0.0480  0.0000498
#> 5 12178   0.0946  q_75     0.0947  0.0000579
#> 6 12455  13.2     max     13.2     0        

# Return observations closest to the five number summary of ln_wages
wages %>%
  keys_near(key = id,
            var = ln_wages)

#> Error in min(x, na.rm = na.rm, ...): invalid 'type' (closure) of argument

# Specify your own list of summaries
l_ranges <- list(min = b_min,
                 range_diff = b_range_diff,
                 max = b_max,
                 iqr = b_iqr)

wages %>%
  key_slope(formula = ln_wages ~ xp) %>%
  keys_near(key = id,
              var = .slope_xp,
              funs = l_ranges)

#> Warning: Outer names are only allowed for unnamed scalar atomic inputs

#> # A tibble: 4 x 5
#>      id .slope_xp stat       stat_value stat_diff
#>   <int>     <dbl> <fct>           <dbl>     <dbl>
#> 1  7918    -4.58  min            -4.58   0       
#> 2 12455    13.2   range_diff     17.8    4.58    
#> 3 12455    13.2   max            13.2    0       
#> 4  9357     0.103 iqr             0.102  0.000139