Return keys nearest to a given statistics or summary.

Return keys nearest to a given statistics or summary.

# S3 method for tbl_ts
keys_near(.data, var, top_n = 1, funs = l_five_num, stat_as_factor = TRUE, ...)

Arguments

.data
tsibble

var
variable to summarise

top_n
top number of closest observations to return - default is 1, which will also return ties.

funs
named list of functions to summarise by. Default is a given list of the five number summary, l_five_num.

stat_as_factor
coerce stat variable into a factor? Default is TRUE.

...
extra arguments to pass to mutate_at when performing the summary as given by funs.

Examples

               
# Return observations closest to the five number summary of ln_wages
wages %>%
  keys_near(var = ln_wages)

#> # A tibble: 71 x 5
#>       id ln_wages stat  stat_value stat_diff
#>    <int>    <dbl> <fct>      <dbl>     <dbl>
#>  1   223     2.14 q_75        2.14         0
#>  2   304     1.84 med         1.84         0
#>  3   470     1.84 med         1.84         0
#>  4   537     1.84 med         1.84         0
#>  5   630     1.84 med         1.84         0
#>  6   700     1.59 q_25        1.59         0
#>  7   735     4.30 max         4.30         0
#>  8   767     1.84 med         1.84         0
#>  9   871     1.84 med         1.84         0
#> 10  1150     1.84 med         1.84         0
#> # … with 61 more rows