Calculating longnostics - statistics of interest for each individual

Calculate statistical summaries for each individual, referred to as a longnostics, a portmanteau of longitudinal and cognostic - a term coined by Tukey which is itself a portmanteau of "cognitive" and "diagnostic". These longnostics all start with l_, and calculate a statistic for each individual in the data, who can be identified with some id, with the statistic being calculated for a specific variable. The current longnostics implemented are:

l_n_obs() Number of observations

l_min() Minimum

l_max() Maximum

l_mean() Mean

l_diff() Lagged difference (by default, the first order difference)

l_q1() First quartile

l_median() Median value

l_q3() Third quartile

l_sd() Standard deviation

l_slope() Slope and intercept (given some linear model formula)

l_mean(data, id, var)

l_sd(data, id, var)

l_max(data, id, var)

l_min(data, id, var)

l_median(data, id, var)

l_q1(data, id, var)

l_q3(data, id, var)

l_diff(data, id, var, lag = 1)

l_n_obs(data, id)

l_slope(data, id, formula)

Arguments

data
data.frame to explore

id
vector of ids to define which values belong to which individual

var
vector of values for individuals, needs to match the id vector

lag
the lag to use, default to 1 (used with l_diff)

formula
character, a formula representing the slope of interest (used in l_slope)

Value

dataframe with column id and l_statistic. For example, l_mean returns the columns id specified, and l_mean.

See also

Examples

l_mean(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id l_mean
#>    <int>  <dbl>
#>  1    31   1.75
#>  2    36   2.33
#>  3    53   1.89
#>  4   122   2.17
#>  5   134   2.48
#>  6   145   1.76
#>  7   155   2.17
#>  8   173   1.93
#>  9   206   2.27
#> 10   207   2.11
#> # … with 878 more rows

l_sd(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id  l_sd
#>    <int> <dbl>
#>  1    31 0.277
#>  2    36 0.387
#>  3    53 0.562
#>  4   122 0.574
#>  5   134 0.321
#>  6   145 0.185
#>  7   155 0.362
#>  8   173 0.274
#>  9   206 0.228
#> 10   207 0.327
#> # … with 878 more rows

l_max(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id l_max
#>    <int> <dbl>
#>  1    31  2.13
#>  2    36  2.93
#>  3    53  3.24
#>  4   122  2.92
#>  5   134  2.93
#>  6   145  2.04
#>  7   155  2.64
#>  8   173  2.34
#>  9   206  2.48
#> 10   207  2.66
#> # … with 878 more rows

l_min(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id l_min
#>    <int> <dbl>
#>  1    31 1.43 
#>  2    36 1.80 
#>  3    53 1.54 
#>  4   122 0.763
#>  5   134 2.00 
#>  6   145 1.48 
#>  7   155 1.54 
#>  8   173 1.56 
#>  9   206 2.03 
#> 10   207 1.58 
#> # … with 878 more rows

l_median(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id l_median
#>    <int>    <dbl>
#>  1    31     1.73
#>  2    36     2.32
#>  3    53     1.71
#>  4   122     2.19
#>  5   134     2.36
#>  6   145     1.77
#>  7   155     2.22
#>  8   173     2.00
#>  9   206     2.30
#> 10   207     2.15
#> # … with 878 more rows

l_q1(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id  l_q1
#>    <int> <dbl>
#>  1    31  1.49
#>  2    36  2.05
#>  3    53  1.59
#>  4   122  2.12
#>  5   134  2.29
#>  6   145  1.60
#>  7   155  1.93
#>  8   173  1.76
#>  9   206  2.16
#> 10   207  1.90
#> # … with 878 more rows

l_diff(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id l_diff_1
#>    <int>    <dbl>
#>  1    31    0.377
#>  2    36    1.11 
#>  3    53    1.70 
#>  4   122    1.68 
#>  5   134    0.368
#>  6   145    0.289
#>  7   155    0.627
#>  8   173    0.319
#>  9   206    0.269
#> 10   207    0.399
#> # … with 878 more rows

l_diff(wages, id, lnw, lag = 2)

#> # A tibble: 888 x 2
#>       id l_diff_2
#>    <int>    <dbl>
#>  1    31    0.462
#>  2    36    0.775
#>  3    53    1.48 
#>  4   122    1.42 
#>  5   134    0.508
#>  6   145    0.440
#>  7   155    0.785
#>  8   173    0.417
#>  9   206    0.454
#> 10   207    0.482
#> # … with 878 more rows

l_q3(wages, id, lnw)

#> # A tibble: 888 x 2
#>       id  l_q3
#>    <int> <dbl>
#>  1    31  1.97
#>  2    36  2.54
#>  3    53  1.88
#>  4   122  2.42
#>  5   134  2.77
#>  6   145  1.88
#>  7   155  2.40
#>  8   173  2.02
#>  9   206  2.39
#> 10   207  2.24
#> # … with 878 more rows

l_n_obs(wages, id)

#> # A tibble: 888 x 2
#>       id l_n_obs
#>    <int>   <int>
#>  1    31       8
#>  2    36      10
#>  3    53       8
#>  4   122      10
#>  5   134      12
#>  6   145       9
#>  7   155      11
#>  8   173       6
#>  9   206       3
#> 10   207      11
#> # … with 878 more rows

l_slope(wages, id, lnw~exper)

#> # A tibble: 888 x 3
#>       id l_intercept l_slope_exper
#>    <int>       <dbl>         <dbl>
#>  1    31        1.41        0.101 
#>  2    36        2.04        0.0588
#>  3    53        2.29       -0.358 
#>  4   122        1.93        0.0374
#>  5   134        2.03        0.0831
#>  6   145        1.59        0.0469
#>  7   155        1.66        0.0867
#>  8   173        1.61        0.100 
#>  9   206        1.73        0.180 
#> 10   207        1.62        0.0884
#> # … with 878 more rows