Calculate all longnostics

This function calculates all longnostics:

l_diff()

l_max()

l_mean()

l_median()

l_min()

l_n_obs()

l_q1()

l_q3()

l_sd()

l_slope()

longnostic_all(data, id, var, formula, lag = 1)

Arguments

data
data.frame to explore

id
vector of ids to define which values belong to which individual

var
vector of values for individuals, needs to match the id vector

formula
character, a formula representing the slope of interest (used in l_slope)

lag
the lag to use, default to 1 (used with l_diff)

Value

a dataframe with longnostics

Examples

longnostic_all(wages, id = id, var = lnw, formula = lnw~exper)

#> # A tibble: 888 x 12
#>       id l_diff_1 l_max l_mean l_median l_min l_n_obs  l_q1  l_q3  l_sd
#>    <int>    <dbl> <dbl>  <dbl>    <dbl> <dbl>   <int> <dbl> <dbl> <dbl>
#>  1    31    0.377  2.13   1.75     1.73 1.43        8  1.49  1.97 0.277
#>  2    36    1.11   2.93   2.33     2.32 1.80       10  2.05  2.54 0.387
#>  3    53    1.70   3.24   1.89     1.71 1.54        8  1.59  1.88 0.562
#>  4   122    1.68   2.92   2.17     2.19 0.763      10  2.12  2.42 0.574
#>  5   134    0.368  2.93   2.48     2.36 2.00       12  2.29  2.77 0.321
#>  6   145    0.289  2.04   1.76     1.77 1.48        9  1.60  1.88 0.185
#>  7   155    0.627  2.64   2.17     2.22 1.54       11  1.93  2.40 0.362
#>  8   173    0.319  2.34   1.93     2.00 1.56        6  1.76  2.02 0.274
#>  9   206    0.269  2.48   2.27     2.30 2.03        3  2.16  2.39 0.228
#> 10   207    0.399  2.66   2.11     2.15 1.58       11  1.90  2.24 0.327
#> # … with 878 more rows, and 2 more variables: l_intercept <dbl>,
#> #   l_slope_exper <dbl>