Student data from 2000-2018 PISA OECD data

A subset of PISA data, containing scores and other information from the triennial testing of 15 year olds around the globe. Original data available from https://www.oecd.org/pisa/data/. Data derived from https://github.com/ropenscilabs/learningtower.

pisa

Format

A tibble of the following variables

year the year of measurement

country the three letter country code

school_id The unique school identification number

student_id The student identification number

gender recorded gender - 1 female or 2 male or missing

math Simulated score in mathematics

read Simulated score in reading

science Simulated score in science

stu_wgt The final survey weight score for the student score

Examples

pisa

#> # A tibble: 405,571 x 9
#>     year country school_id student_id gender  math science  read stu_wgt
#>    <int> <fct>       <int>      <int>  <int> <dbl>   <dbl> <dbl>   <dbl>
#>  1  2000 AUS          1006          1      1   NA     469.  526     5.60
#>  2  2000 AUS          1006          2      2   NA     607.  559.    5.60
#>  3  2000 AUS          1006          3      2   NA     335.  383.    5.60
#>  4  2000 AUS          1006          4      1   NA     576   595     5.60
#>  5  2000 AUS          1006          7      2  550.    585.  497.    5.60
#>  6  2000 AUS          1006          9      2   NA     516.  532.    5.60
#>  7  2000 AUS          1006         10      2   NA     514.  575.    5.60
#>  8  2000 AUS          1006         12      2  377     384.  325.    5.60
#>  9  2000 AUS          1006         13      2   NA     458.  452.    5.60
#> 10  2000 AUS          1006         14      2  588.    499   507.    5.60
#> # … with 405,561 more rows