brolgar: Browse Over Longitudinal Data Graphically and Analytically in R

Provides a framework of tools to summarise, visualise, and explore longitudinal data. It builds upon the tidy time series data frames used in the 'tsibble' package, and is designed to integrate within the 'tidyverse', and 'tidyverts' (for time series) ecosystems. The methods implemented include calculating features for understanding longitudinal data, including calculating summary statistics such as quantiles, medians, and numeric ranges, sampling individual series, identifying individual series representative of a group, and extending the facet system in 'ggplot2' to facilitate exploration of samples of data. These methods are fully described in the paper "brolgar: An R package to Browse Over Longitudinal Data Graphically and Analytically in R", Nicholas Tierney, Dianne Cook, Tania Prvan (2020) arXiv:2012.01619.

Details

brolgar stands for: BRowse over Longitudinal data Graphically and Analytically in R.

See also

Author

Maintainer: Nicholas Tierney nicholas.tierney@gmail.com (ORCID)

Authors:

Di Cook dicook@monash.edu (ORCID)

Tania Prvan tania.prvan@mq.edu.au

Other contributors:

Stuart Lee [contributor]

Earo Wang [contributor]