Filter by the number of observations for an id.

When exploring longitudinal data it can be useful to filter by the number of observations to help with certain exploratory data analysis.

filter_n_obs(data, id, filter)

Arguments

data
data.frame to explore

id
vector of ids to define which values belong to which individual

filter
an expression you want to filter by, where number of observations is referred in as l_n_obs. For example (l_n_obs > 10) would filter those observations with greater than 10 observations.

Value

dataframe filtered by the number of observations, with an additional column l_n_obs containing the number of observations for each id.

Examples

wages %>% filter_n_obs(id = id, filter = l_n_obs > 10)

#> # A tibble: 1,105 x 10
#>       id l_n_obs   lnw exper   ged postexp black hispanic   hgc uerate
#>    <int>   <int> <dbl> <dbl> <int>   <dbl> <int>    <int> <int>  <dbl>
#>  1   134      12  2.00 0.192     0    0        0        0    11   5.8 
#>  2   134      12  2.10 0.972     0    0        0        0    11   5.49
#>  3   134      12  2.32 1.93      0    0        0        0    11   6.09
#>  4   134      12  2.39 2.95      0    0        0        0    11   8.4 
#>  5   134      12  2.26 3.92      1    0        0        0    11   8.1 
#>  6   134      12  2.33 4.64      1    0.72     0        0    11   6.39
#>  7   134      12  2.30 5.64      1    1.72     0        0    11   5.89
#>  8   134      12  2.67 6.75      1    2.84     0        0    11   5.09
#>  9   134      12  2.81 8.25      1    4.34     0        0    11   2.8 
#> 10   134      12  2.93 9.02      1    5.11     0        0    11   3.19
#> # … with 1,095 more rows

wages %>% filter_n_obs(id = id, filter = l_n_obs == 2)

#> # A tibble: 78 x 10
#>       id l_n_obs   lnw exper   ged postexp black hispanic   hgc uerate
#>    <int>   <int> <dbl> <dbl> <int>   <dbl> <int>    <int> <int>  <dbl>
#>  1   518       2 1.27  0.525     1   0.525     1        0     8  10.7 
#>  2   518       2 1.61  1.93      1   1.93      1        0     8  10.7 
#>  3  1687       2 1.66  0.688     0   0         1        0    10   5.39
#>  4  1687       2 1.26  0.825     0   0         1        0    10   5.3 
#>  5  2178       2 1.52  0.923     0   0         0        0     8   8.1 
#>  6  2178       2 1.75  1.96      0   0         0        0     8  11.1 
#>  7  2456       2 0.898 4         0   0         0        0    11  12.1 
#>  8  2456       2 1.22  5         0   0         0        0    11  12   
#>  9  2612       2 1.42  0.875     1   0.875     0        0     8   7.73
#> 10  2612       2 1.57  1.14      1   1.14      0        0     8   7.73
#> # … with 68 more rows