Return keys nearest to a given statistics or summary.

Return keys nearest to a given statistics or summary.

# S3 method for tbl_ts
keys_near(.data, var, top_n = 1, funs = l_five_num, stat_as_factor = TRUE, ...)

Arguments

.data
data.frame

var
variable to summarise

top_n
top number of closest observations to return - default is 1, which will also return ties.

funs
named list of functions to summarise by. Default is a given list of the five number summary, l_five_num.

stat_as_factor
coerce stat variable into a factor? Default is TRUE.

...
extra arguments to pass to mutate_at when performing the summary as given by funs.

Examples

# Return observations closest to the five number summary of ln_wages
wages %>%
  keys_near(var = ln_wages)

#> # A tibble: 71 x 5
#>       id ln_wages stat  stat_value stat_diff
#>    <int>    <dbl> <fct>      <dbl>     <dbl>
#>  1  4385    0.708 min        0.708         0
#>  2   700    1.59  q_25       1.59          0
#>  3  4025    1.59  q_25       1.59          0
#>  4  4072    1.59  q_25       1.59          0
#>  5  4221    1.59  q_25       1.59          0
#>  6  6035    1.59  q_25       1.59          0
#>  7  6746    1.59  q_25       1.59          0
#>  8  7963    1.59  q_25       1.59          0
#>  9  8979    1.59  q_25       1.59          0
#> 10  9316    1.59  q_25       1.59          0
#> # … with 61 more rows