Return keys nearest to a given statistics or summary.

Return keys nearest to a given statistics or summary.

Usage

# S3 method for tbl_ts
keys_near(.data, var, top_n = 1, funs = l_five_num, stat_as_factor = TRUE, ...)

Arguments

.data

tsibble

var

variable to summarise

top_n

top number of closest observations to return - default is 1, which will also return ties.

funs

named list of functions to summarise by. Default is a given list of the five number summary, l_five_num.

stat_as_factor

coerce stat variable into a factor? Default is TRUE.

...

extra arguments to pass to mutate_at when performing the summary as given by funs.

Examples

               
# Return observations closest to the five number summary of height_cm
heights %>%
  keys_near(var = height_cm)
#> # A tibble: 18 × 5
#>    country          height_cm stat  stat_value stat_diff
#>    <chr>                <dbl> <fct>      <dbl>     <dbl>
#>  1 Denmark               183. max         183.   0      
#>  2 Ethiopia              167. med         167.   0.00900
#>  3 Ghana                 164. q_25        164.   0      
#>  4 Hungary               164. q_25        164.   0      
#>  5 Italy                 164. q_25        164.   0      
#>  6 Italy                 164. q_25        164.   0      
#>  7 Liberia               167. med         167.   0.00900
#>  8 Morocco               170. q_75        170.   0.00392
#>  9 Mozambique            164. q_25        164.   0      
#> 10 Mozambique            164. q_25        164.   0      
#> 11 Pakistan              164. q_25        164.   0      
#> 12 Papua New Guinea      152. min         152.   0      
#> 13 Romania               164. q_25        164.   0      
#> 14 Romania               164. q_25        164.   0      
#> 15 Sierra Leone          164. q_25        164.   0      
#> 16 Slovakia              164. q_25        164.   0      
#> 17 Ukraine               164. q_25        164.   0      
#> 18 Ukraine               164. q_25        164.   0